jobs/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
tqdm==4.66.1